Minor pathfinder changes and base for spawning towers and units

+++ src/se/liu/ida/oscth887oskth878/tddc69/project/Util/Point.java
@@ -1,18 +1,19 @@
 package se.liu.ida.oscth887oskth878.tddc69.project.Util;
 
 /**
- * Created with IntelliJ IDEA.
- * User: Oscar
- * Date: 2013-09-15
- * Time: 20:48
+ * @version 1.0
+ * @author: Oskar Therén
+ * @author: Oscar Thunberg
+ * @since: 29/09/2013
  */
 public class Point {
-    public float x, y;
+    public int x, y;
 
-    public Point(float x, float y) {
+    public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
+
 
     @Override
     public String toString() {
